change for legacy reprocessing of 2011 data.
In short:
I have started with the systematic code for 2012 and replaced parts by
the 2011 version:
I kept the overall structure and all sub categories for relative and
absolute.

absolute:
 I have set Stat, Scale and Mpfbias to their 2011 values. I kept flavor
mapping at 0 and changed the reference for spr and frag to 100 GeV.

relative:
 I have set JER, FSR and Stat to their 2011 values. However, I kept
RelativePt at its 2012 value as I had the fear that we might miss
something otherwise as also the single pion response and the
fragmentation is taken from 2012 and I kept the flavor treatment as it
is in 2012.

PU: I have changed the NPV to 8, but kept the 2012 uncertainty code. I
assumed that the new MC for 2011 has the same bug as the 2012 MC with
53x. As we will use the 2012 L1 for data and MC, it might be safer to
keep all PU uncertainties.

Time: I have used half the 2011 time dependence. Dominik's plots
indicate a reduction in the time dependence, but we also added an ad-hoc
residual as the mean response shifted around 2.7 for the legacy
processing.

Flavor: I kept the 2012 numbers. I like our new approach and hopefully
the change from 8 to 7 TeV does not change much.